feat(ui): adds seperate JN drawer, keeps original
