+++ app/src/components/navigation-bar.tsx
@@ -45,6 +45,7 @@ import { useTheme } from "next-themes";
 import { FeatureFlags, hasFeatureFlag } from "@/util/feature-flags";
 import { useOrganizationContext } from "@/hooks/organization-context-provider/use-organizational-context";
 import { Trans } from "react-i18next";
+import JNDrawer from "./HomePage/JNDrawer";
 
 function countryFromLanguage(language: string) {
   return language == "en" ? "us" : language;
@@ -99,6 +100,11 @@ export function NavigationBar({
     history.replaceState(null, "", newPath);
   };
 
+  // get pathname
+  const pathname = usePathname();
+  const fullPath = pathname.replace(/^\/[A-Za-z]+/, "");
+  console.log("fullPath", fullPath);
+
   // Checks if language is set in cookie and updates URL if not
   useEffect(() => {
     if (typeof window !== "undefined") {
@@ -127,7 +133,6 @@ export function NavigationBar({
   const currentInventoryId =
     inventoryIdFromParam ?? userInfo?.defaultInventoryId;
   const router = useRouter();
-  const pathname = usePathname();
   const inventoryStub = inventoryIdFromParam ?? currentInventoryId;
   const dashboardPath = `/${lng}/${inventoryStub === null ? "" : inventoryStub}`;
   const { setTheme } = useTheme();
@@ -491,13 +496,28 @@ export function NavigationBar({
             )}
           </Box>
         </Box>
-        <ProjectDrawer
-          lng={lng}
-          currentInventoryId={currentInventoryId as string}
-          isOpen={isDrawerOpen}
-          onClose={() => setIsDrawerOpen(false)}
-          onOpenChange={({ open }) => setIsDrawerOpen(open)}
-        />
+        {/* JN Drawer */}
+        {/* Should be shown if JN is enabled and url has /cities*/}
+        {fullPath.includes("/cities") && (
+          <JNDrawer
+            lng={lng}
+            currentInventoryId={currentInventoryId as string}
+            isOpen={isDrawerOpen}
+            onClose={() => setIsDrawerOpen(false)}
+            onOpenChange={({ open }) => setIsDrawerOpen(open)}
+          />
+        )}
+        {/* TODO: Remove project drawer and replace with JN drawer after JN is live */}
+        {/* Project Drawer */}
+        {!fullPath.includes("/cities") && (
+          <ProjectDrawer
+            lng={lng}
+            currentInventoryId={currentInventoryId as string}
+            isOpen={isDrawerOpen}
+            onClose={() => setIsDrawerOpen(false)}
+            onOpenChange={({ open }) => setIsDrawerOpen(open)}
+          />
+        )}
       </Box>
       {isFrozen && !isPublic && !isAuth && (
         <Box py={2} px={16} bg="sentiment.warningDefault" w="full" zIndex={50}>
