+++ app/src/i18n/locales/de/dashboard.json
@@ -185,6 +185,11 @@
   "search-by-city-or-project": "Nach Stadt oder Projekt suchen",
   "add-new-city": "Neue Stadt hinzufügen",
   "city": "Stadt",
+  "ghg-inventories": "THG-Inventare",
+  "climate-risks": "Klimarisiken",
+  "cc-inventories": "CC-Inventare",
+  "actions": "Aktionen",
+  "finance-readiness": "Finanzbereitschaft",
   "select-project-to-invite-collaborators": "Wählen Sie das Projekt aus, um Mitarbeiter einzuladen",
   "project-to-share": "Wählen Sie aus, welches Projekt Sie teilen möchten",
   "upgrade-plan-city-limit-heading": "Müssen Sie Ihren Tarif aktualisieren?",
