+++ app/src/i18n/locales/en/dashboard.json
@@ -187,6 +187,11 @@
   "search-by-city-or-project": "Search by city or project",
   "add-new-city": "Add new city",
   "city": "City",
+  "ghg-inventories": "GHG Inventories",
+  "climate-risks": "Climate Risks",
+  "cc-inventories": "CC Inventories",
+  "actions": "Actions",
+  "finance-readiness": "Finance Readiness",
   "select-project-to-invite-collaborators": "Select project to invite collaborators",
   "project-to-share": "Select which project you want to share",
   "upgrade-plan-city-limit-heading": "Need to Upgrade your plan?",
