+++ app/src/i18n/locales/es/dashboard.json
@@ -188,6 +188,11 @@
   "search-by-city-or-project": "Buscar por ciudad o proyecto",
   "add-new-city": "Agregar nueva ciudad",
   "city": "Ciudad",
+  "ghg-inventories": "Inventarios de GEI",
+  "climate-risks": "Riesgos climáticos",
+  "cc-inventories": "Inventarios CC",
+  "actions": "Acciones",
+  "finance-readiness": "Preparación Financiera",
   "select-project-to-invite-collaborators": "Selecciona el proyecto para invitar a colaboradores",
   "project-to-share": "Selecciona qué proyecto quieres compartir",
   "upgrade-plan-city-limit-heading": "¿Necesitas actualizar tu plan?",
