+++ app/src/i18n/locales/fr/dashboard.json
@@ -185,6 +185,11 @@
   "search-by-city-or-project": "Rechercher par ville ou projet",
   "add-new-city": "Ajouter une nouvelle ville",
   "city": "Ville",
+  "ghg-inventories": "Inventaires GES",
+  "climate-risks": "Risques climatiques",
+  "cc-inventories": "Inventaires CC",
+  "actions": "Actions",
+  "finance-readiness": "Préparation Financière",
   "select-project-to-invite-collaborators": "Sélectionnez le projet pour inviter des collaborateurs",
   "project-to-share": "Sélectionnez le projet que vous souhaitez partager",
   "upgrade-plan-city-limit-heading": "Besoin de mettre à niveau votre plan ?",
