+++ app/src/i18n/locales/pt/dashboard.json
@@ -186,6 +186,11 @@
   "search-by-city-or-project": "Pesquisar por cidade ou projeto",
   "add-new-city": "Adicionar nova cidade",
   "city": "Cidade",
+  "ghg-inventories": "Inventários de GEE",
+  "climate-risks": "Riscos climáticos",
+  "cc-inventories": "Inventários CC",
+  "actions": "Ações",
+  "finance-readiness": "Preparação Financeira",
   "select-project-to-invite-collaborators": "Selecione o projeto para convidar colaboradores",
   "project-to-share": "Selecione qual projeto você deseja compartilhar",
   "upgrade-plan-city-limit-heading": "Precisa atualizar o seu plano?",
